~ | norm
